backend/requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
python-dotenv
gunicorn
Flask-Migrate
PyJWT
Flask-CORS
flask_jwt_extended